Move createNewFile so that its next to Create since only Create uses it.
Rename openWithAppend and openWithWrite to openForAppend and openForWrite to better communicate what they are doing.

Change computeAndUpdateHash to be synchronous rather than called with go. This makes the logic easier. Also have it update the Hasher state for the AccessedFileState.

Remove Sequence from AccessedFileState, this just complicates the logic trying to do things in the background that can be done synchronously.